Commit 5 - Executable code
Left -
Remove tf flags, save models at checkpoints

+++ model/gcn/init.py
@@ -3,7 +3,7 @@ import numpy as np
 
 def uniform(shape, scale=0.05, name=None):
     """Uniform init."""
-    initial = torch.FloatTensor(shape).uniform_(-scale, scale)
+    initial = torch.DoubleTensor(shape).uniform_(-scale, scale)
    # initial = tf.random_uniform(shape, minval=-scale, maxval=scale, dtype=tf.float32)
     return initial
 
@@ -12,7 +12,7 @@ def weight_variable_glorot(input_dim, output_dim, name=""):
     """Glorot & Bengio (AISTATS 2010) init."""
     init_range = np.sqrt(6.0 / (input_dim + output_dim))
     
-    initial = torch.FloatTensor(input_dim, output_dim).uniform_(-init_range, init_range)
+    initial = torch.DoubleTensor(input_dim, output_dim).uniform_(-init_range, init_range)
  #   initial = tf.random_uniform([input_dim, output_dim], minval=-init_range,
   #                              maxval=init_range, dtype=tf.float32)
     return initial
